IntをBigIntに置き換えた

+++ src/main/scala/Parser.scala
@@ -23,5 +23,5 @@ object Parse extends RegexParsers {
 		"/" ^^ {op => (l, r) => Div(l, r)}
 	)
 	def parenthesis: Parser[Base] = "(" ~ expr ~ ")" ^^ {case "(" ~ exp ~ ")" => exp}
-	def number: Parser[Base] = """(\+|-)?\d+""".r ^^ {s => Int(s.toInt)}
+	def number: Parser[Base] = """(\+|-)?\d+""".r ^^ {s => Int(BigInt(s))}
 }
